src/interfaces/api.py
"""LlamaHome REST API interface.

This module implements the API endpoints described in docs/API.md, enabling
communication between components and interaction with the Llama 3.3 model.

@see docs/API.md
"""

import logging
from pathlib import Path
from typing import Any, Dict, List, Optional, Union

from fastapi import FastAPI, HTTPException, WebSocket
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel

from utils.log_manager import LogManager, LogTemplates

logger = LogManager(LogTemplates.SYSTEM_STARTUP).get_logger(__name__)

# Initialize FastAPI app
app = FastAPI(
    title="LlamaHome API",
    description="REST API for LlamaHome system",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# API Models
class ModelConfig(BaseModel):
    """Model configuration parameters."""
    model_path: str
    temperature: Optional[float] = 0.7
    max_tokens: Optional[int] = 100
    top_p: Optional[float] = 0.9
    frequency_penalty: Optional[float] = 0.0
    presence_penalty: Optional[float] = 0.0

class ProcessRequest(BaseModel):
    """Prompt processing request."""
    prompt: str
    model: Optional[str] = "llama3.3"
    temperature: Optional[float] = 0.7
    max_tokens: Optional[int] = 100
    top_p: Optional[float] = 0.9
    frequency_penalty: Optional[float] = 0.0
    presence_penalty: Optional[float] = 0.0

# API Endpoints
@app.post("/api/load_model")
async def load_model(config: ModelConfig) -> Dict[str, Any]:
    """Load a specific Llama 3.3 model into memory."""
    try:
        # Implementation
        return {"status": "success", "message": "Model loaded successfully"}
    except Exception as e:
        logger.error(f"Failed to load model: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/process_prompt")
async def process_prompt(request: ProcessRequest) -> Dict[str, Any]:
    """Process a user prompt and generate a response."""
    try:
        # Implementation
        return {
            "response": "Generated response",
            "tokens_used": 0,
            "processing_time": 0.0,
            "model_used": request.model
        }
    except Exception as e:
        logger.error(f"Failed to process prompt: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/api/health_check")
async def health_check() -> Dict[str, Any]:
    """Check API and model health status."""
    try:
        return {
            "status": "success",
            "message": "API is healthy",
            "model_loaded": True,
            "uptime": 0
        }
    except Exception as e:
        logger.error(f"Health check failed: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

@app.websocket("/api/ws")
async def websocket_endpoint(websocket: WebSocket):
    """WebSocket endpoint for real-time updates."""
    try:
        await websocket.accept()
        while True:
            data = await websocket.receive_text()
            await websocket.send_text(f"Message received: {data}")
    except Exception as e:
        logger.error(f"WebSocket error: {str(e)}")
        await websocket.close()
